+++ tilings/strategies/assumption_splitting.py
@@ -114,11 +114,10 @@ class Split(Constructor):
     ):
         raise NotImplementedError
 
-    def __eq__(self, obj: object) -> bool:
+    def equiv(
+        self, other: "Constructor", data: Optional[object] = None
+    ) -> Tuple[bool, Optional[object]]:
         raise NotImplementedError("Required for bijections")
-
-    def __hash__(self) -> int:
-        raise NotImplementedError("Required for bijection search")
 
 
 class SplittingStrategy(Strategy[Tiling, GriddedPerm]):
@@ -164,6 +163,19 @@ class SplittingStrategy(Strategy[Tiling, GriddedPerm]):
     @staticmethod
     def is_two_way(comb_class: Tiling):
         return False
+
+    @staticmethod
+    def is_reversible(comb_class: Tiling):
+        return False
+
+    def shifts(
+        self, comb_class: Tiling, children: Optional[Tuple[Tiling, ...]]
+    ) -> Tuple[int, ...]:
+        if children is None:
+            children = self.decomposition_function(comb_class)
+            if children is None:
+                raise StrategyDoesNotApply
+        return (0,)
 
     def decomposition_function(self, comb_class: Tiling) -> Optional[Tuple[Tiling]]:
         if not comb_class.assumptions: